requirements.txt
django
mozilla-django-oidc
django-allauth
djangorestframework
requests
pyjwt